feat/upload  Changes in errorcode + New model ImageData + put the logic for upload_imafe in order app
