+++ orders/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import OrderImageModel, OrderModel, OrderOffer
+from .models import OrderImageModel, OrderModel, OrderOffer, ImageData
 
 
 @admin.register(OrderModel)
@@ -10,6 +10,11 @@ class OrderModelAdmin(admin.ModelAdmin):
 @admin.register(OrderImageModel)
 class OrderImageModelAdmin(admin.ModelAdmin):
     list_display = ["id", "order_id"]
+
+
+@admin.register(ImageData)
+class ImageDataModelAdmin(admin.ModelAdmin):
+    list_display = ["id", "user_account", "order_id"]
 
 
 @admin.register(OrderOffer)
