+++ products/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 
 from .views import (AnswerListAPIView, CardModelAPIView, ImageResponseAPIView,
                     CategoryModelListAPIView, QuestionsModelListAPIView, CreateOrderAnswers)
-from . import views
 
 
 urlpatterns = [
@@ -16,5 +15,4 @@ urlpatterns = [
     path("products/step_4/create_order", CreateOrderAnswers),
     path("products/responses/image", ImageResponseAPIView.as_view()),
 
-    path("products/image_upload_order/", views.upload_image_order, name='upload-image-order'),
 ]
