+++ utils/image_work.py
@@ -1,1 +1,12 @@
 """ логика для работы с изображениями """
+
+
+class ImageWork(object):
+    def __int__(self):
+        self.image_size = None
+        self.name = None
+
+    def resize_preview(self, image):
+        pass
+
+
